files reorg

+++ sources/visualize_map.c
@@ -6,7 +6,7 @@
 /*   By: tzanchi <tzanchi@student.42berlin.de>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/07/25 17:11:34 by tzanchi           #+#    #+#             */
-/*   Updated: 2023/08/07 15:49:24 by tzanchi          ###   ########.fr       */
+/*   Updated: 2023/08/08 15:04:25 by tzanchi          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -56,21 +56,21 @@ void	resize_map(t_fdf **fdf)
 
 void	put_cross(t_vect_3 v, mlx_image_t *img, int colour)
 {
-	if ((int)v.x >= 0 && (int)v.x <= WIDTH
-		&& (int)v.y >= 0 && (int)v.y <= HEIGTH)
-		mlx_put_pixel(img, (int)v.x, (int)v.y, colour);
-	if ((int)v.x >= 1 && (int)v.x <= WIDTH + 1
-		&& (int)v.y >= 0 && (int)v.y <= HEIGTH)
-		mlx_put_pixel(img, (int)v.x - 1, (int)v.y, colour);
-	if ((int)v.x >= -1 && (int)v.x <= WIDTH - 1
-		&& (int)v.y >= 0 && (int)v.y <= HEIGTH)
-		mlx_put_pixel(img, (int)v.x + 1, (int)v.y, colour);
-	if ((int)v.x >= 0 && (int)v.x <= WIDTH
-		&& (int)v.y >= 1 && (int)v.y <= HEIGTH + 1)
-		mlx_put_pixel(img, (int)v.x, (int)v.y - 1, colour);
-	if ((int)v.x >= 0 && (int)v.x <= WIDTH
-		&& (int)v.y >= -1 && (int)v.y <= HEIGTH - 1)
-		mlx_put_pixel(img, (int)v.x, (int)v.y + 1, colour);
+	if (v.x >= 0 && v.x <= WIDTH
+		&& v.y >= 0 && v.y <= HEIGTH)
+		mlx_put_pixel(img, v.x, v.y, colour);
+	if (v.x >= 1 && v.x <= WIDTH + 1
+		&& v.y >= 0 && v.y <= HEIGTH)
+		mlx_put_pixel(img, v.x - 1, v.y, colour);
+	if (v.x >= -1 && v.x <= WIDTH - 1
+		&& v.y >= 0 && v.y <= HEIGTH)
+		mlx_put_pixel(img, v.x + 1, v.y, colour);
+	if (v.x >= 0 && v.x <= WIDTH
+		&& v.y >= 1 && v.y <= HEIGTH + 1)
+		mlx_put_pixel(img, v.x, v.y - 1, colour);
+	if (v.x >= 0 && v.x <= WIDTH
+		&& v.y >= -1 && v.y <= HEIGTH - 1)
+		mlx_put_pixel(img, v.x, v.y + 1, colour);
 }
 
 /*Loops through the list pointed at by 'fdf' and connects all nodes to the
